requirements.txt
 transformers 
 accelerate 
 bitsandbytes
 pipeline
 pymupdf 
 pdfplumber 
 langchain 
 chromadb 
 llama-index 
 neo4j 
 sentence-transformers
 typing
 flask
 DateTime
 pillow 
 requests
 google-cloud-vision 
 pymupdf

